% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate Simulation Results Across Replications}
\usage{
aggregate_results(all_results)
}
\arguments{
\item{all_results}{A list of results from multiple replications.}
}
\value{
A list containing aggregated theta, item, and correlation summaries.
}
\description{
This function aggregates results from multiple replications of simulations, computing
metrics such as bias and RMSE for both theta estimates and item parameters.
}
\examples{
all_results <- list(run_condition(conditions[1, ], 2), run_condition(conditions[2, ], 2))
aggregate_results(all_results)
}
