% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_item_par.R
\name{generate_item_par}
\alias{generate_item_par}
\title{Generate Fixed Item Parameters}
\usage{
generate_item_par(n_items)
}
\arguments{
\item{n_items}{Number of items to simulate.}
}
\value{
A list containing discrimination (\code{a}) and difficulty (\code{b}) parameters.
}
\description{
This function generates fixed item parameters for simulation.
}
\examples{
generate_item_par(12)
}
